drm/nouveau/timer: restore the time on resume
This can be useful if some parts of Nouveau try to calculate the time between
two events. Without this patch, the time difference would be negative in the
case where the computer is suspended/resumed between two events.
This patch should fix fan speed probing when done while suspending/resuming.

Solve this by saving the current time before suspending and by restoring it
on resume.

Signed-off-by: Martin Peres <martin.peres@labri.fr>
Tested-by: Martin Peres <martin.peres@labri.fr>